+++ DatadogObjc/Sources/DDURLSessionDelegate+objc.swift
@@ -9,40 +9,39 @@ import DatadogCore
 import DatadogInternal
 
 @objc
-open class DDNSURLSessionDelegate: NSObject, URLSessionTaskDelegate, URLSessionDataDelegate, __URLSessionDelegateProviding {
-    var swiftDelegate: DDURLSessionDelegate
-    public var ddURLSessionDelegate: DatadogURLSessionDelegate {
-        return swiftDelegate
-    }
-
+@available(*, deprecated, message: "Use `URLSessionInstrumentation.enable(with:)` instead.")
+open class DDNSURLSessionDelegate: NSObject, URLSessionTaskDelegate, URLSessionDataDelegate {
     @objc
     override public init() {
-        swiftDelegate = DDURLSessionDelegate()
+        URLSessionInstrumentation.enable(
+            with: .init(
+                delegateClass: Self.self
+            ),
+            in: CoreRegistry.default
+        )
+        super.init()
     }
 
     @objc
     public init(additionalFirstPartyHostsWithHeaderTypes: [String: Set<DDTracingHeaderType>]) {
-        swiftDelegate = DDURLSessionDelegate(
-            additionalFirstPartyHostsWithHeaderTypes: additionalFirstPartyHostsWithHeaderTypes.mapValues { tracingHeaderTypes in
-                return Set(tracingHeaderTypes.map { $0.swiftType })
-            }
+        URLSessionInstrumentation.enable(
+            with: .init(
+                delegateClass: Self.self,
+                firstPartyHostsTracing: .traceWithHeaders(hostsWithHeaders: additionalFirstPartyHostsWithHeaderTypes.mapValues { tracingHeaderTypes in
+                    return Set(tracingHeaderTypes.map { $0.swiftType })
+                })
+            ),
+            in: CoreRegistry.default
         )
+        super.init()
     }
 
     @objc
-    public init(additionalFirstPartyHosts: Set<String>) {
-        swiftDelegate = DDURLSessionDelegate(additionalFirstPartyHosts: additionalFirstPartyHosts)
-    }
-
-    open func urlSession(_ session: URLSession, task: URLSessionTask, didCompleteWithError error: Error?) {
-        swiftDelegate.urlSession(session, task: task, didCompleteWithError: error)
-    }
-
-    open func urlSession(_ session: URLSession, task: URLSessionTask, didFinishCollecting metrics: URLSessionTaskMetrics) {
-        swiftDelegate.urlSession(session, task: task, didFinishCollecting: metrics)
-    }
-
-    open func urlSession(_ session: URLSession, dataTask: URLSessionDataTask, didReceive data: Data) {
-        swiftDelegate.urlSession(session, dataTask: dataTask, didReceive: data)
+    public convenience init(additionalFirstPartyHosts: Set<String>) {
+        self.init(
+            additionalFirstPartyHostsWithHeaderTypes: additionalFirstPartyHosts.reduce(into: [:], { partialResult, host in
+                partialResult[host] = [.datadog]
+            })
+        )
     }
 }