+++ TestUtilities/Helpers/XCTestCase.swift
@@ -47,4 +47,20 @@ extension XCTestCase {
             }
         }
     }
+
+    @available(iOS 13.0, tvOS 13.0, *)
+    public func dd_fulfillment(
+        for expectations: [XCTestExpectation],
+        timeout seconds: TimeInterval = .infinity,
+        enforceOrder enforceOrderOfFulfillment: Bool = false) async {
+#if compiler(>=5.8)
+            await fulfillment(of: expectations, timeout: seconds, enforceOrder: enforceOrderOfFulfillment)
+#else
+        wait(
+            for: expectations,
+            timeout: seconds,
+            enforceOrder: enforceOrderOfFulfillment
+        )
+#endif
+    }
 }